dev-requirements.txt
# Development and Testing Requirements
# For code quality, linting, type checking, and testing

# Code quality and linting
pylint>=3.0.0
flake8>=7.0.0
black>=24.0.0
isort>=5.13.0
ruff>=0.2.0

# Type checking
mypy>=1.8.0
types-PyYAML>=6.0.0
types-requests>=2.31.0
types-setuptools>=69.0.0

# Testing and coverage
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
coverage>=7.4.0
pandas>=2.0.0
psycopg[binary]>=3.1.0
requests-mock>=1.11.0
matplotlib>=3.8.0

# Schema validation
pandera>=0.18.0
pydantic>=2.5.0

# Pre-commit hooks
pre-commit>=3.6.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0
